+++ other/changelog.md
@@ -3,3 +3,18 @@ All notable changes to this project will be documented in this file. The format 
 
 
 ## [Unreleased]
+
+## [0.0.001] - 12.10.2018
+
+### Pushed the code for the first time, pre documentation.
+
+## [0.0.002] - 31.10.2018
+
+### Performance
+	• Created hashmap of sub sprites, takes a sprite name as a string argument as returns image of the subsprite with that name.
+
+### General
+	• Refactored variable names to be consistent project wide.
+	• Included dynamic checking for shadow blocks.
+	• 'Brick' tiles are now updatable in real time with mouse clicks.
+	• Animations for decaying 'Brick' type tiles has been included.